fix(noir-proof-batch): pR fixes incorporated
