+++ package.json
@@ -21,7 +21,6 @@
         "test:contracts-noir": "yarn workspace semaphore-contracts-noir test:coverage",
         "test:circuits": "yarn workspace @semaphore-protocol/circuits test",
         "test:circuits-noir": "jest packages/circuits-noir --forceExit; cd ./packages/circuits-noir && nargo test",
-        "test:noir-batch": "jest packages/noir-proof-batch",
         "lint": "concurrently -c auto -g -n eslint,solhint \"eslint . --ext .js,.ts,.tsx\" \"yarn workspace semaphore-contracts lint\"",
         "format": "concurrently -c auto -g -n prettier,embark \"prettier -c .\" \"yarn workspace semaphore-docs format\"",
         "format:write": "concurrently -c auto -g -n prettier,embark \"prettier -w .\" \"yarn workspace semaphore-docs format:write\"",
