+++ packages/noir-proof-batch/src/vks.ts
@@ -1,0 +1,263 @@
+// vk for batching 2 leaves circuit
+export const batch2LeavesCircuitVk = [
+    "0x0000000000000000000000000000000000000000000000000000000000200000",
+    "0x0000000000000000000000000000000000000000000000000000000000000010",
+    "0x0000000000000000000000000000000000000000000000000000000000000001",
+    "0x0000000000000000000000000000000000000000000000000000000000000001",
+    "0x0000000000000000000000000000000000000000000000000000000000000000",
+    "0x0000000000000000000000000000000000000000000000000000000000000001",
+    "0x0000000000000000000000000000000000000000000000000000000000000002",
+    "0x0000000000000000000000000000000000000000000000000000000000000003",
+    "0x0000000000000000000000000000000000000000000000000000000000000004",
+    "0x0000000000000000000000000000000000000000000000000000000000000005",
+    "0x0000000000000000000000000000000000000000000000000000000000000006",
+    "0x0000000000000000000000000000000000000000000000000000000000000007",
+    "0x0000000000000000000000000000000000000000000000000000000000000008",
+    "0x0000000000000000000000000000000000000000000000000000000000000009",
+    "0x000000000000000000000000000000000000000000000000000000000000000a",
+    "0x000000000000000000000000000000000000000000000000000000000000000b",
+    "0x000000000000000000000000000000000000000000000000000000000000000c",
+    "0x000000000000000000000000000000000000000000000000000000000000000d",
+    "0x000000000000000000000000000000000000000000000000000000000000000e",
+    "0x000000000000000000000000000000000000000000000000000000000000000f",
+    "0x00000000000000000000000000000025d90d90d95607ee29179ae5c722ae0c8f",
+    "0x00000000000000000000000000000000000d9c6041d10416fed659315d27aee2",
+    "0x00000000000000000000000000000049a1673d9c80d3059520ad6f977e2c9bc5",
+    "0x00000000000000000000000000000000002dd52f3dea230276bc3edef3130d91",
+    "0x0000000000000000000000000000004e6a79cdc860550bc4e497bd0681041c18",
+    "0x0000000000000000000000000000000000047c2c64988e4e9f80c223517fc8d7",
+    "0x000000000000000000000000000000d90f731c058187f4daff5768c89ebf0ea0",
+    "0x00000000000000000000000000000000002a2986112eb9cb946252c2d3d5b71c",
+    "0x00000000000000000000000000000000e48d95071907f820a1b305ec17e84b37",
+    "0x000000000000000000000000000000000009f3a1cc177d9ee735ffe1bc9b6d74",
+    "0x00000000000000000000000000000061b35ccf60653e98c1c2c2df971002f98c",
+    "0x0000000000000000000000000000000000149a3b203a2ad1333e3472e592ba03",
+    "0x000000000000000000000000000000e50ec01bc6364d0200e909b211ecc131b4",
+    "0x00000000000000000000000000000000002bfbe9e9264f79c7302943970efba8",
+    "0x0000000000000000000000000000003d00176c87547ab6e16df4408a38955e42",
+    "0x000000000000000000000000000000000029708adc33ff55955b4c667c828e0a",
+    "0x000000000000000000000000000000d409ae42b51b3f255eb9dbb0bc60be7f12",
+    "0x00000000000000000000000000000000002fc601b1726373e455a3032ad7d705",
+    "0x000000000000000000000000000000baa7a2b8902f14448b08c2cdf1048fdbd7",
+    "0x00000000000000000000000000000000000a1ebe57220868ebba8e33d49c01e4",
+    "0x000000000000000000000000000000b51c323bd9513bdc02302ee79f34b2537e",
+    "0x0000000000000000000000000000000000120653d53577b54a2562fd49f34c6b",
+    "0x0000000000000000000000000000000ebcda0a505c27be3d0f8184608d051038",
+    "0x000000000000000000000000000000000002b014e691431ff73bd89bdd653531",
+    "0x000000000000000000000000000000ece661d9ddf8d6ce71e63ee377b120f70f",
+    "0x000000000000000000000000000000000017c6d9d50e48678a2ac344538de4c7",
+    "0x000000000000000000000000000000ad54bcdd8f21bebc775e9dfb36b9a73d45",
+    "0x000000000000000000000000000000000019c51b736e4c5a7d8380246160d19a",
+    "0x000000000000000000000000000000899c524eb8f3dbf98513c080690e50912c",
+    "0x000000000000000000000000000000000011bbffbe11c18b7cf3b93f70b7b468",
+    "0x000000000000000000000000000000fc6b7f3254e2440315f20cdec9c3b99bdb",
+    "0x0000000000000000000000000000000000204776e1e90171068f2c6c0d1fbb90",
+    "0x000000000000000000000000000000c2ca9daf2371f1a4f487b80dcb46c31216",
+    "0x000000000000000000000000000000000020c82983472aff106a884ee18de4fe",
+    "0x000000000000000000000000000000851b1c22bcc040095fc1ba4a5dd7e6ab1e",
+    "0x0000000000000000000000000000000000033b339f8aff3dd85ab91fdd116bce",
+    "0x0000000000000000000000000000004e1d633aed64174e18424c9d229cafb00b",
+    "0x0000000000000000000000000000000000228bd9a70d6c95555bb52b1ceb4e33",
+    "0x000000000000000000000000000000b180a0949218e63f7229209d8f82cd7db6",
+    "0x00000000000000000000000000000000002338ab6892811bb85b99d4655f2e56",
+    "0x000000000000000000000000000000dc0963b49e7afcbbfb2a822346b5d122bc",
+    "0x000000000000000000000000000000000005ee52fe5faf71df5c4f2c682ab4b8",
+    "0x000000000000000000000000000000f3388d608bb2384d0a21bd1dae125e949c",
+    "0x00000000000000000000000000000000002316296eec45d16306b02ab2b239c5",
+    "0x0000000000000000000000000000007e45e57e96b09269a8009ebec3243e7b2d",
+    "0x00000000000000000000000000000000002588d486346c14ac23bcb23458ed83",
+    "0x0000000000000000000000000000000fd712e9431ed9219342a3735c1be2ff36",
+    "0x00000000000000000000000000000000001d581d403a123b54af4b16d47c6d0a",
+    "0x00000000000000000000000000000011355ffe679b0eab1edb3d75deae5ce8ae",
+    "0x000000000000000000000000000000000019e5d8c0c7f50ffb971c7a561d6402",
+    "0x0000000000000000000000000000006d838e0a96bc15212dea031b642721ba25",
+    "0x000000000000000000000000000000000026bedae660132627a30fd4a5d7b024",
+    "0x000000000000000000000000000000a707b6f71cb5e947d1fd3a50b9aa917974",
+    "0x000000000000000000000000000000000008fed61b7bb8e67d08ec8cdb282c3f",
+    "0x000000000000000000000000000000491c59408057bcfa3ee1dead9ddc185ec3",
+    "0x0000000000000000000000000000000000276dc18b846e1896387e6ea1a21d98",
+    "0x0000000000000000000000000000006429622380d15c65862c904d54c1575ca7",
+    "0x000000000000000000000000000000000015fedf772114d5a5a068bb85119cbe",
+    "0x000000000000000000000000000000ad7d3460547249f491090141c25e762d7c",
+    "0x0000000000000000000000000000000000117ebf85e995a61480a3c56f3798df",
+    "0x000000000000000000000000000000c239db1ca2314504bb24d32f0bf03da5d0",
+    "0x00000000000000000000000000000000001439bc2b1cf977bad90ae102932964",
+    "0x000000000000000000000000000000729ef23a7851e836264006f331d68d33e3",
+    "0x0000000000000000000000000000000000085e7b2a7598e031ab1eb621ee2d06",
+    "0x00000000000000000000000000000000599953cdec0f0cef565e2220af47f468",
+    "0x00000000000000000000000000000000000572288b2d16303af3ba1d6325c587",
+    "0x0000000000000000000000000000007e5832369c157fd9ad8eb236fe1329ea6c",
+    "0x0000000000000000000000000000000000048abbb777fd2644ee404ddad8b6be",
+    "0x000000000000000000000000000000d2a29dca40812643f3906e5cf40881b5bb",
+    "0x00000000000000000000000000000000000687428fa06ada305f5a769acec663",
+    "0x000000000000000000000000000000d3963254ef71a3ede9d50ff76aafe3e2fd",
+    "0x0000000000000000000000000000000000098274da1a1d257171004adf94be02",
+    "0x000000000000000000000000000000bb83625dea36a7b41fc81fa83c29cb6408",
+    "0x00000000000000000000000000000000000431a6cf7f40877bdfb2a28b66c2b6",
+    "0x0000000000000000000000000000002a52a082bf57795cbc5053d49e8dc24982",
+    "0x00000000000000000000000000000000001a125646963aa95ec4c54ec66c26e6",
+    "0x000000000000000000000000000000cbc5fb38232c26b94378c424fde735f2bd",
+    "0x00000000000000000000000000000000001fac614c9a0d82bdfa04393b2dfbd7",
+    "0x0000000000000000000000000000003e780c4b7541e005adb424d760aa618888",
+    "0x0000000000000000000000000000000000174ce5ef9cd5c6cc1ae881b600e9ff",
+    "0x00000000000000000000000000000062cf7b0a70b09ac6c78d1996442e445f94",
+    "0x00000000000000000000000000000000000987021c7b62ef76501a6829a5fc81",
+    "0x0000000000000000000000000000007c9cb4bde3f1db6aaff2bca01b440cfbbc",
+    "0x00000000000000000000000000000000000ff82e91bb8df91f4d89b7db75e1ef",
+    "0x000000000000000000000000000000243cf3e0cfba606d27d5999f4927ff92b3",
+    "0x00000000000000000000000000000000001f1156b93b4396e0dac3bd312fdc94",
+    "0x000000000000000000000000000000419d9961d76a65ed28914ca5cc3ffd2433",
+    "0x0000000000000000000000000000000000116a7935196d39ea9178a285c53a6b",
+    "0x000000000000000000000000000000f369409c76a0245d4f389193b554c30065",
+    "0x000000000000000000000000000000000023aebc5efc1d0e6d03030b242308fd",
+    "0x0000000000000000000000000000000cbfc1214af084c189478e34dc04c77419",
+    "0x000000000000000000000000000000000019f38f8e7cf18f375d75db06fca92a",
+    "0x0000000000000000000000000000000a771b5bbb501c75790a1a4e2906931045",
+    "0x000000000000000000000000000000000015642d62fc17d119ba4afb77ab424e",
+    "0x000000000000000000000000000000a5c1396036397af54a729801cc1c37d4e2",
+    "0x000000000000000000000000000000000021cea98314ec6efc5f8f1f648f42a7",
+    "0x000000000000000000000000000000dd5b4a42e59fe1e447cad24659049d13a7",
+    "0x00000000000000000000000000000000001f3bd0ebf0709ac30745d0dafb183c",
+    "0x00000000000000000000000000000016c2efd51d298fee5fce4355fc26890195",
+    "0x000000000000000000000000000000000005900180ddd1cec6e340c70c9bff6f",
+    "0x0000000000000000000000000000000000000000000000000000000000000001",
+    "0x0000000000000000000000000000000000000000000000000000000000000000",
+    "0x0000000000000000000000000000000000000000000000000000000000000002",
+    "0x0000000000000000000000000000000000000000000000000000000000000000",
+    "0x0000000000000000000000000000000dcdc1cd46a781f41c20ef020a82cecfcf",
+    "0x00000000000000000000000000000000000037bf30f89056701d394772e180c3",
+    "0x00000000000000000000000000000018dbc2a9a804a5a763c7aec8584666c63e",
+    "0x000000000000000000000000000000000029958d9570523421e135078c07236a"
+]
+
+// vk for batching 2 nodes circuit
+export const batch2NodesCircuitVk = [
+    "0x0000000000000000000000000000000000000000000000000000000000200000",
+    "0x0000000000000000000000000000000000000000000000000000000000000010",
+    "0x0000000000000000000000000000000000000000000000000000000000000001",
+    "0x0000000000000000000000000000000000000000000000000000000000000001",
+    "0x0000000000000000000000000000000000000000000000000000000000000000",
+    "0x0000000000000000000000000000000000000000000000000000000000000001",
+    "0x0000000000000000000000000000000000000000000000000000000000000002",
+    "0x0000000000000000000000000000000000000000000000000000000000000003",
+    "0x0000000000000000000000000000000000000000000000000000000000000004",
+    "0x0000000000000000000000000000000000000000000000000000000000000005",
+    "0x0000000000000000000000000000000000000000000000000000000000000006",
+    "0x0000000000000000000000000000000000000000000000000000000000000007",
+    "0x0000000000000000000000000000000000000000000000000000000000000008",
+    "0x0000000000000000000000000000000000000000000000000000000000000009",
+    "0x000000000000000000000000000000000000000000000000000000000000000a",
+    "0x000000000000000000000000000000000000000000000000000000000000000b",
+    "0x000000000000000000000000000000000000000000000000000000000000000c",
+    "0x000000000000000000000000000000000000000000000000000000000000000d",
+    "0x000000000000000000000000000000000000000000000000000000000000000e",
+    "0x000000000000000000000000000000000000000000000000000000000000000f",
+    "0x000000000000000000000000000000853793114cc97784c0662f7fb49093b15a",
+    "0x000000000000000000000000000000000005db220d982083e5d8fbb5964ab426",
+    "0x00000000000000000000000000000041efd0de6aff4ff3c7895cb7101c7a31a1",
+    "0x000000000000000000000000000000000025c91f18c4d5a4e1b44eac89a8373a",
+    "0x0000000000000000000000000000006cdcb45f8e2fc34ad9a8dbb2c2088acca4",
+    "0x0000000000000000000000000000000000212e0cf40c1e663429ad834887b6bc",
+    "0x00000000000000000000000000000021a19734b786b054f8b7798fe84075052e",
+    "0x000000000000000000000000000000000015baba9110cd7bf6157e312b203437",
+    "0x0000000000000000000000000000002db4b48aa7aebcf12a4eff376ca91907e7",
+    "0x0000000000000000000000000000000000224dffa2aaeaedc6b6149fa7a4cb2d",
+    "0x000000000000000000000000000000da6e92ce2205843e532b31c3346d748b33",
+    "0x000000000000000000000000000000000004cb894b27e063af560b46c25aa713",
+    "0x000000000000000000000000000000b76ac3dd8f68ad0138d95e5bd666877d0c",
+    "0x00000000000000000000000000000000000a9a3d8a12af2e2b80e73f7b4fb1a2",
+    "0x00000000000000000000000000000041493046ff7a4d6d8fcbf8208e4eb2233a",
+    "0x00000000000000000000000000000000001065db107324d016164ccdd9a93728",
+    "0x00000000000000000000000000000046bf3ee4402294368f6b5336044336ddf3",
+    "0x00000000000000000000000000000000001b247d3663a44c03704ea7e08f608e",
+    "0x0000000000000000000000000000000ca97c4255791a0c1dad0ff4eba3440bb3",
+    "0x00000000000000000000000000000000000935ba6c9da3223fa5cb39217ce065",
+    "0x000000000000000000000000000000afcd68c0b0e3097570bd9ab388244628b5",
+    "0x00000000000000000000000000000000000b542cc196cc8b0eb8b86042ea807f",
+    "0x00000000000000000000000000000017f1f77ab3edcf7902cb5260f5dfe22952",
+    "0x00000000000000000000000000000000000fd19ff666baf07b4e749b472f7b75",
+    "0x000000000000000000000000000000ece661d9ddf8d6ce71e63ee377b120f70f",
+    "0x000000000000000000000000000000000017c6d9d50e48678a2ac344538de4c7",
+    "0x000000000000000000000000000000ad54bcdd8f21bebc775e9dfb36b9a73d45",
+    "0x000000000000000000000000000000000019c51b736e4c5a7d8380246160d19a",
+    "0x0000000000000000000000000000003f3b0fca9997b0581d08bea04ddc0752e0",
+    "0x0000000000000000000000000000000000212f1f6c8a813551f1e52ec48c11ec",
+    "0x0000000000000000000000000000007b6fbb3d28ff034c91c1e68a7eac726185",
+    "0x00000000000000000000000000000000000f637990bdbd5adbd912afd325ab84",
+    "0x000000000000000000000000000000796ee4dcfaac96d5a23126decc7912a0d0",
+    "0x00000000000000000000000000000000002dd763eabd10d073f4b249f0ac9f69",
+    "0x000000000000000000000000000000fc4c964e1df2e7329b832525e379351d8a",
+    "0x000000000000000000000000000000000024f49b460ae83eee1d323044f1ca9f",
+    "0x0000000000000000000000000000000f2f33f7d03d31143e9565c827c5586ca8",
+    "0x000000000000000000000000000000000028054cb80059e066f2f4efa92071cd",
+    "0x000000000000000000000000000000eebf6061224abe73ddcf1207efe2ce49e1",
+    "0x000000000000000000000000000000000010e11931417ce0a6612890b8ee982b",
+    "0x000000000000000000000000000000d0ac32ebd9a08132a26175aa89a8f7e639",
+    "0x00000000000000000000000000000000002d23ee42c21be324ec226f9c7a84fd",
+    "0x000000000000000000000000000000c84f9e50b6b3b146333bffbbdd880cf7b6",
+    "0x00000000000000000000000000000000002db48d08acdceae4beb0df6ffac301",
+    "0x0000000000000000000000000000009f85538bd81adcf3a21d179c351cd34ccc",
+    "0x0000000000000000000000000000000000137a790e4e5fc4e6e3e2c291590498",
+    "0x000000000000000000000000000000925336e705dd8a85153954b7f7992a8a0c",
+    "0x00000000000000000000000000000000000a6b574aa29f7771b627bd5a5c2b78",
+    "0x000000000000000000000000000000fa3e2c735c34d138fe59cbbf7836431d97",
+    "0x0000000000000000000000000000000000286835edb3d8659f60690e4b4b44c4",
+    "0x000000000000000000000000000000ada5a7ad87f2e52ff76a305a2bf0790a70",
+    "0x00000000000000000000000000000000001525b6fba23e02e5e78e1f41b028ae",
+    "0x000000000000000000000000000000b13e818265bad80a923ca0eaf38161467f",
+    "0x0000000000000000000000000000000000010843239d7b7a2e9c66625f2913fe",
+    "0x000000000000000000000000000000fe70ed9011f8bc88590eaa4f5158a4ee57",
+    "0x0000000000000000000000000000000000141ce316b7ca52edd89b46aae19ffe",
+    "0x00000000000000000000000000000079b8a22dd373d4848de905c1dd528320ba",
+    "0x000000000000000000000000000000000021723f776efeb964ff91ac76e63c4b",
+    "0x0000000000000000000000000000007d3ec2c8ab38ce9e53252c9bf3c6266305",
+    "0x00000000000000000000000000000000001930025d49aa27db7b52aad16d9c30",
+    "0x000000000000000000000000000000458657ae33a7da34c092c3837ebd0014d0",
+    "0x00000000000000000000000000000000000248ab05faae9973fa1717292832ad",
+    "0x0000000000000000000000000000001c43e9a2129968d95fb701d316fc1e7767",
+    "0x00000000000000000000000000000000002e3b09e1da086a912a286a4c6a4dff",
+    "0x0000000000000000000000000000009d03ae2f9e8f5741b54398002d36f628be",
+    "0x00000000000000000000000000000000002acb44e717fff904dc65845d6571bb",
+    "0x00000000000000000000000000000036d9aac2a8abda95dc86934cbdae4edf77",
+    "0x000000000000000000000000000000000017914e40c463428fa4e860bdf2be5a",
+    "0x0000000000000000000000000000003fcab131869553ab7269019a852a6e19d5",
+    "0x0000000000000000000000000000000000111337dbd2c760ab3c16b736c3e096",
+    "0x000000000000000000000000000000d05cf70547518af7295c0b1665734363c6",
+    "0x00000000000000000000000000000000000d8f188d46d4662c6e450723693dab",
+    "0x000000000000000000000000000000b2f67a8ab2d56c649c23ead379cae6701a",
+    "0x0000000000000000000000000000000000109474fbac03e39a12123247d537b5",
+    "0x000000000000000000000000000000656225dbea2977aeaaf8a580fbe3535f83",
+    "0x000000000000000000000000000000000003bb08f5ad597bcc784d8866cebfda",
+    "0x0000000000000000000000000000005c71ae752a6849552b1cfa147243cf4bb3",
+    "0x000000000000000000000000000000000009bb85ad242bfd649cede6fd276c49",
+    "0x000000000000000000000000000000d4f0f8d94462a97dd586a70b32aa12bc51",
+    "0x0000000000000000000000000000000000063a3fef5f97c2b4291837fb473d6e",
+    "0x0000000000000000000000000000007b53f89c7f953bcc775e2bc18209dc9679",
+    "0x00000000000000000000000000000000002ab7067253088d36c661bd99c9bd36",
+    "0x0000000000000000000000000000000fe074d48f6c99836c769fe43d8ac16fd6",
+    "0x000000000000000000000000000000000006f57a8492c43a8ab8d61b828e0001",
+    "0x000000000000000000000000000000243cf3e0cfba606d27d5999f4927ff92b3",
+    "0x00000000000000000000000000000000001f1156b93b4396e0dac3bd312fdc94",
+    "0x000000000000000000000000000000419d9961d76a65ed28914ca5cc3ffd2433",
+    "0x0000000000000000000000000000000000116a7935196d39ea9178a285c53a6b",
+    "0x000000000000000000000000000000f369409c76a0245d4f389193b554c30065",
+    "0x000000000000000000000000000000000023aebc5efc1d0e6d03030b242308fd",
+    "0x0000000000000000000000000000000cbfc1214af084c189478e34dc04c77419",
+    "0x000000000000000000000000000000000019f38f8e7cf18f375d75db06fca92a",
+    "0x0000000000000000000000000000000a771b5bbb501c75790a1a4e2906931045",
+    "0x000000000000000000000000000000000015642d62fc17d119ba4afb77ab424e",
+    "0x000000000000000000000000000000a5c1396036397af54a729801cc1c37d4e2",
+    "0x000000000000000000000000000000000021cea98314ec6efc5f8f1f648f42a7",
+    "0x000000000000000000000000000000dd5b4a42e59fe1e447cad24659049d13a7",
+    "0x00000000000000000000000000000000001f3bd0ebf0709ac30745d0dafb183c",
+    "0x00000000000000000000000000000016c2efd51d298fee5fce4355fc26890195",
+    "0x000000000000000000000000000000000005900180ddd1cec6e340c70c9bff6f",
+    "0x0000000000000000000000000000000000000000000000000000000000000001",
+    "0x0000000000000000000000000000000000000000000000000000000000000000",
+    "0x0000000000000000000000000000000000000000000000000000000000000002",
+    "0x0000000000000000000000000000000000000000000000000000000000000000",
+    "0x000000000000000000000000000000c4b6fb9150ed11685a09038ea35070642d",
+    "0x0000000000000000000000000000000000007280c5169a0e1f859a500138817e",
+    "0x000000000000000000000000000000f7151c97526c14e4531a6b1778adccedf5",
+    "0x000000000000000000000000000000000016658259e029bb766fe0d828342f9e"
+]
